Add files via upload
These are the HTML and CSS files that run the website I use as a cover letter to showcase my skills as a front-end web developer. I manage a in progress column and a done column in my projects to show what changes i plan to making and what I've already done with it here on my github. As I update the website, I'll update these files. To properly display these they need to be in a folder by themselves. If i sent them to you as part of my resume they they probably already came that way but if you wish to download them just to check it out then keep that in mind.